+++ leituras/models.py
@@ -13,8 +13,19 @@ class Leitura(models.Model):
     mes = models.IntegerField()
     ano = models.IntegerField()
 
-    leitura_agua = models.FloatField(blank=True, null=True)
-    leitura_gas = models.FloatField(blank=True, null=True)
+    leitura_agua = models.DecimalField(
+        max_digits=10,
+        decimal_places=2,
+        blank=True,
+        null=True,
+    )
+
+    leitura_gas = models.DecimalField(
+        max_digits=10,
+        decimal_places=2,
+        blank=True,
+        null=True,
+    )
 
     data_registro = models.DateTimeField(auto_now_add=True)
